update README to reflect `cloning` repo info
The README is confusing as it simply says run `cargo run` as if it is a binary and can be executed anywhere. Added info on cloning the repo and actually running it from the root of the repo. Though it looks like a trivial change, current explanation is very confusing for new rust users.